requirements.txt
binance-api-node
talib
flask
